関数やメソッドを切り出し①
