+++ js/main.js
@@ -1,19 +1,12 @@
 const submitButton = document.querySelector('.js_addTodo_submit');
-const toggle = document.querySelector('.js_todoTask_toggle');
 const ul = document.querySelector('.js_todoTask_list');
 const input = document.querySelector('.js_addTodo_input');
 import { Todo } from './todo.js';
+import checkInput from './checkInput.js';
+import dropdown from './dropdown.js';
 
-const checkInput = () => {
-  if (/^\S/.test(input.value)) {
-    submitButton.disabled = false;
-    submitButton.style.opacity = 1;
-    return;
-  }
-  submitButton.disabled = true;
-  submitButton.style.opacity = 0.5;
-};
 checkInput();
+dropdown();
 
 input.addEventListener('input', () => {
   checkInput();
@@ -29,19 +22,10 @@ submitButton.addEventListener('click', () => {
   input.value = '';
   checkInput();
 
+  
+
   const label = document.querySelector('.js_todoTask_label');
   label.addEventListener('change', (e) => {
     newTodo.updateTodo(e);
   });
-});
-
-//ドロップダウン、closedクラスの付け外し
-toggle.addEventListener('click', (e) => {
-  if (e.target.classList.contains('closed')) {
-    e.target.classList.remove('closed');
-    ul.style.display = 'block';
-    return;
-  }
-  e.target.classList.add('closed');
-  ul.style.display = 'none';
 });
